move cli to its own directory
Makes things more uniform with other parts of conda.
