+++ bin/conda-build
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 import sys
-from conda_build.main_build import main
+
+from conda_build.cli.main_build import main
 
 sys.exit(main())
