+++ bin/conda-convert
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 import sys
-from conda_build.main_convert import main
+
+from conda_build.cli.main_convert import main
 
 sys.exit(main())
