+++ bin/conda-develop
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 import sys
-from conda_build.main_develop import main
+
+from conda_build.cli.main_develop import main
 
 sys.exit(main())
