+++ bin/conda-index
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 import sys
-from conda_build.main_index import main
+
+from conda_build.cli.main_index import main
 
 sys.exit(main())
