+++ bin/conda-inspect
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 import sys
-from conda_build.main_inspect import main
+
+from conda_build.cli.main_inspect import main
 
 sys.exit(main())
