+++ bin/conda-metapackage
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 import sys
-from conda_build.main_metapackage import main
+
+from conda_build.cli.main_metapackage import main
 
 sys.exit(main())
