+++ bin/conda-pipbuild
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 import sys
-from conda_build.main_pipbuild import main
+
+from conda_build.cli.main_pipbuild import main
 
 sys.exit(main())
