+++ bin/conda-sign
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 import sys
-from conda_build.main_sign import main
+
+from conda_build.cli.main_sign import main
 
 sys.exit(main())
