+++ bin/conda-skeleton
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 import sys
-from conda_build.main_skeleton import main
+
+from conda_build.cli.main_skeleton import main
 
 sys.exit(main())
