+++ conda_build/__main__.py
@@ -1,4 +1,5 @@
 import sys
-from conda_build.main_build import main
+
+from conda_build.cli.main_build import main
 
 sys.exit(main())
